Global Search Cleanup
