+++ src/Core/GlobalSearh/Config/ProCoSys/types.ts
